Prime Count C++ new #88
